Begin tracking post-3.3-pre3 development with git.
Thanks to Luiz Fernando N. Capitulino for his mail, suggesting that it might be
really good to begin tracking development with a proper SCM. This is especially
important now that we're trying to actively encourage greater involvement in
development once more. I'm looking for more patches and git trees to merge!