Subject: rmmod: Changes exit() return value.
In the current implementation if rmmod fails, it will return the last
error it got.

This is not very good because it's not consistent and if the error is
negative, we'll have not standard return error codes from rmmod (eg, 255)

This patch changes that behaivor by always returning 1 if at least one
of the modules to remove failed.

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@mandriva.com.br>
Signed-off-by: Jon Masters <jcm@jonmasters.org>