Make module-init-tools build with ANSI C 89 compiler
From: Petter Reinholdtsen <pere@hungry.com>

The source for module-init-tools version 3.3-pre4-2 fail to build with
an ANSI C 89 compiler, because variables are defined after a statement
in one place of the code.  This patch solve the problem.  Please
include it to make the source easier to build in sub-optimal
environments.