don't warn noisely if loading multiple module aliases
Another patch that has been sitting in Fedora for a while. Basically, this
happened because systems doing root LUKS would get a whiny error about a
lack of VIA padlock if the hardware was not present. So we don't get
loud by default. I think this helps other distros so upstreaming,
however, I'll be redoing this patch to make it better.

Signed-off-by: Jon Masters <jcm@jonmasters.org>