add versioning to the binary module files
We need to have a version attached to the binary trie in case we ever
need to change the format in some fashion. So this patch does that, and
also changes the MODULE_MAGIC to 0xB007F457 (change the 'A' to a '4')
so that we can notice that we have the second version of the files.

Signed-off-by: Jon Masters <jcm@jonmasters.org>