depmod: Fix modules.dep truncation
if a wrong module name (or a nonexistent module) is passed to depmod it
will truncate the modules.dep file.

The problem is that even if the modules list is empty, depmod will try
to do its job. This is wrong, it should do nothing if no modules have
been found.

Signed-off-by: Luiz Fernando N. Capitulino <lcapitulino@mandriva.com.br>
Signed-off-by: Jon Masters <jcm@jonmasters.org>