add module ordering support for binary indexes
Tejun Heo added support to prioritize modules based
on their position in modules.order, in the same way
built-in drivers are prioritized by link order.
Unfortunately this didn't work by the time it was
applied, because we added fast binary indexes,
which did not preserve ordering.

This required a rethink of the index code.  Overall
it is now a few lines shorter and slightly easier
to understand.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>