depmod: Allow modules outside the normal module directory
We used to allow this.  But we switched to relative paths
and it stopped working without any warning at depmod time.

This can only happen if depmod is run with a list of modules.
Probably no-one is affected.  But failing silently is bad,
and once we add the check for this case it is trivial to
support it.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>