Fix underscores() corner case (and optimise)
The pattern "[[]-" matches the string "[-".
underscores() should convert the pattern to "[[]_".

The current code gets this wrong, leaving it as "[[]-".
It seems to think patterns can include nested square brackets.
This shouldn't affect anyone, but it does makes it a bit faster.

Also, we can reduce the number of calls to underscores()
for non-matching aliases.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>