modprobe: remove flock() because it never worked anyway
We used to have a file lock/unlock using flock() and acquiring a "write"
lock on the module file prior to load/unload as a means to acquire an
exclusive access to the module file. Sounds good, right? The problem is
that this will instead block if anyone else is holding the file open in
userspace (as well as the intention of preventing concurrent modprobe).

We could just wait 5 seconds, print warnings, kill the offending other
process, or many other things. But since modprobe already works on a
read-only filesystem (by not taking the lock at all, allowing the
kernel to catch and prevent the concurrent load situation) we do
this in every case instead. If I get complaints, I'll try an
alternative mechanism instead - like a warning, kill, etc.

Signed-off-by: Jon Masters <jcm@jonmasters.org>