modprobe: abort loading module if one of its dependents fails to load
The gist of the patch is that once a dependency has failed, propagate the error
up and stop loading all the modules that are dependent on the failed one.  This
allows us to bail out quicker and not run into issues like 'Unknown symbols'
because the parent modules are never asked to load.

Signed-off-by: Jon Masters <jcm@jonmasters.org>