testsuite: add --valgrind
"valgrindme.sh" has been hanging around for a while without doing
anything. The original idea seems to have been to run under valgrind
automatically. But it is very slow, which discourages one from running
the testsuite. Implement it as an option instead.

In particular, this was useful while writing the binary index code.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>
Signed-off-by: Jon Masters <jcm@jonmasters.org>