elfops: export load_section(), hide load_strings()
load_strings() was planned to replace get_section() + next_string()
e.g in modinfo. However it turns out there are no advantages in style
and few in function. As it won't be used outside elfops, the export
is removed.

load_section() OTOH, will replace get_section() so we export it now.

Signed-off-by: Andreas Robinson <andr345@gmail.com>