doc: make a link from modprobe.d.sgml to modprobe.conf.sgml
The previous hack for modprobe.d.5 happens because 'docbook2man' as shipped
in at least one major distribution does not support '--solinks' to create
automatic links for each of the refentrytitle entries. It will actually only
make links for the last one. This is the other piece - for now, link sources.

Signed-off-by: Jon Masters <jcm@jonmasters.org>