elfops: fix crash on grab_fd() failure
If we call release_elf_file() when grab_fd() fails, it will call
release_file() with a NULL pointer and undefined length.  This can
cause a crash when zlib support is disabled and release_file() is
implemented using munmap().

This is only theoretical at the moment.  However it will allow
grab_elf_file_fd() to be removed straightforwardly, without
creating a more significant bug.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>