modprobe: Ignore custom install commands if module_in_kernel() doesn't work
Custom install commands rely on a working module_in_kernel() to avoid
an infinite fork-loop.  This can fail if /sys/module/ is not available
(i.e. before sysfs is mounted).  For example:

    install snd-pcm /sbin/modprobe --ignore-install snd-pcm && \
	{ /sbin/modprobe --quiet snd-pcm-oss ; : ; }

The snd-pcm-oss module depends on snd-pcm.  If we can't tell that
snd-pcm is already loaded when we load snd-pcm-oss, we end up
running the entire install command again, ad infinitim.

Thanks go to Modestas Vainius for the idea and original patch.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>