modprobe: clean up mit_ignore_loaded flag in rmmod()
handle_module() clears mit_ignore_loaded before calling rmmod(). So we
didn't really want to set it in the first place!  Also, rmmod() uses the
flag to mean something which differs both from the name itself, and the
way it is used in insmod().

Replace the mit_ignore_loaded in rmmod() with a new flag mit_quiet_inuse
(it means not to complain if the module cannot be removed because it is
in use).

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>