modprobe: don't run remove commands for modules which are not loaded
modprobe.d(5) strongly suggests you can use remove commands like this:

    remove snd-pcm /sbin/modprobe -r --ignore-remove snd-pcm-oss && \
	/sbin/modprobe -r snd-pcm

Unfortunately, given that snd-pcm-oss depends on snd-pcm, this causes
an infinite fork-loop.

The solution is an exact parallel of the approach taken for install
commands.  AFAIK no-one uses remove commands, but it's good to make
the insmod() and rmmod() functions consistent.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>