modprobe: Don't assume module absent if no /sys/module/<module>/initstate
/sys/module/<module>/initstate only appeared in 2.6.20.  If it is not
present, we cannot tell whether the module has finished loading
succesfully.  In this case module_in_kernel() should return -1
(undefined).

Take care to to distinguish between "initstate disappeared because the
module was just unloaded" and "initstate is absent but the module is
still present".

"Fork bombing" side effect of this bug is described in [1].
A followup patch will add a fallback to /proc/modules.

1. http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=524940

Thanks to Modestas Vainius <modestas@vainius.eu> for explaining the
problem.

Signed-off-by: Alan Jenkins <alan-jenkins@tuffmail.co.uk>