elfops: remove errfn_t from load_strings
Commit 528db92ab1dd0d75dba415b9f3dc81f5a34773ce added an errfn_t
parameter to elfops_core.c:load_strings. This was for the purpose
of detecting missing terminators at the end of ELF-sections with
strings in them, such as .modinfo.

However, the committer (that'd be me) forgot to add any actual code to
load_strings() and now the errfn_t parameter complicates the error
handling when softdep is used.

This commit removes that parameter and adds a non-fatal warning message.

Signed-off-by: Andreas Robinson <andr345@gmail.com>