depmod: use real regex matching on depmod "overrides" entries
We implement support for matching "override" entries in depmod config
files to determine which module of several alternatives will be used.

The existing implementation is a very poor "wildcard" match that does
not use real regular expressions. This version uses the standard POSIX
extended regular expressions, with backward compatibility to match on
a single "*" kernel version as in the original (broken) design.

Signed-off-by: Jon Masters <jcm@jonmasters.org>