modprobe: refactor read_aliases()
It doesn't care whether we are "removing", and it only parses aliases
(not an entire config).

Enhance readability by splitting it into two functions, for binary
and text files respectively.  This matches the way read_depends()
is written.

Fix some sites to use read_aliases() instead of parse_config_file().