modprobe: re-enable inadvertently disabled options
At some point several short options versions (-S, -D, -R) were inadvertently
disabled. The first is used to override the kernel version, the second is used
to dump dependency information, and the third dumps resolved alias information.

The documentation was updated in the manpages in the previous commit and is
now in sync with how modprobe is actually implemented.

Signed-off-by: Jon Masters <jcm@jonmasters.org>