modprobe: trivial cleanups post-merge of softdep bits
In order to merge in the softdep bits from alan, I wound up resetting
the dev tree and remerging the whole lot. Next time, I'll just review
the patches in another tree and then pull the lot, which is better.
These two trivial differences were in my tree and are ok to keep.

Signed-off-by: Jon Masters <jcm@jonmasters.org>