build: remove explicit call to build modules.dep.bin
There's a known bug with docbook2man in some cases (Fedora) not
following multiple names for a manpage and only generating a
single output file. We use links in the source to ensure
those other versions will be built if possible. A better
fix will be to figure out how to make docbook2man behave.

Signed-off-by: Jon Masters <jcm@jonmasters.org>