doc: properly fix documentation generation
The docbook2man package does not allow output to be written to standard out,
nor does the SGML allow for more than one file to be created. A previous
hack was to use symlinks and hope for the best. Rather than doing this,
retain the symlinks, but explicitly match on them during build and
create symlinks to the output files. Furthermore, clean up the
documentation source files to remove invalid multiple outputs.

Signed-off-by: Jon Masters <jcm@jonmasters.org>